Properly split up tests and time compilation
