+++ Laboratory/C++/test/jazz.cpp
@@ -1,4 +1,4 @@
-#include "../gen/models.hpp"
+#include "../gen/jazz.hpp"
 #include <cstdio>
 #include <map>
 #include <vector>
